requirements_image_editor.txt
# 現象学的画像編集システムの依存関係

# 基本的な画像処理
Pillow>=10.0.0

# 高度な画像処理とコンピュータビジョン
opencv-python>=4.8.0

# 科学計算・数値処理
numpy>=1.24.0
scipy>=1.11.0

# 科学的画像処理（オプション - より高度なフィルターが必要な場合）
scikit-image>=0.21.0

# 既存の依存関係（プロジェクト既存）
openai
python-dotenv